Override ZoteroController to pass user key to CreateSubscriptionJob
Sufia::Arkivo::CreateSubscriptionJob, called in ZoteroController#callback, needs to receive the user's key and not the
user object itself.

Uses the prepend function in Rails to inject the new behavior without overriding the whole controller.
